refactor: move table actions & headers to config folder
