+++ src/components/DataTable/DataTableActions.tsx
@@ -1,55 +1,8 @@
 import { IconButton, Tooltip } from "@material-ui/core";
-import {
-  AttachFile,
-  Cancel,
-  CloudUploadOutlined,
-  Delete,
-  Edit,
-  LibraryAdd,
-} from "@material-ui/icons";
+
+import CONFIG from "config/table/actions";
 
 import { cln } from "helpers/cln";
-
-import tableActions from "constants/tableActions";
-
-const CONFIG = {
-  [tableActions.edit]: {
-    get title() {
-      return tableActions.edit;
-    },
-    icon: <Edit color="primary" />,
-  },
-  [tableActions.delete]: {
-    get title() {
-      return tableActions.delete;
-    },
-    icon: <Delete color="secondary" />,
-  },
-  [tableActions.reject]: {
-    get title() {
-      return tableActions.reject;
-    },
-    icon: <Cancel color="secondary" />,
-  },
-  [tableActions.upload]: {
-    get title() {
-      return tableActions.upload;
-    },
-    icon: <CloudUploadOutlined color="primary" />,
-  },
-  [tableActions.download]: {
-    get title() {
-      return tableActions.download;
-    },
-    icon: <AttachFile color="primary" />,
-  },
-  [tableActions.copy]: {
-    get title() {
-      return tableActions.copy;
-    },
-    icon: <LibraryAdd color="primary" />,
-  },
-};
 
 export interface IDataTableActions {
   actions: {
