+++ src/config/table/headers/index.ts
@@ -1,0 +1,3 @@
+import customer from "./customer";
+
+export { customer as customerHeaders };
