+++ src/helpers/getDisplayedValue.ts
