--- src/helpers/getDisplayedValue/definedHeaders/index.ts
@@ -1,3 +1,0 @@
-import customerHeaders from "./customerHeaders";
-
-export default { customerHeaders };
