+++ src/modules/customerManagement/CustomerEdit/CustomerEdit.tsx
@@ -9,7 +9,6 @@ import FormFooter from "components/FormFooter";
 import ProgressSpinner from "components/ProgressSpinner";
 
 import getDisplayedValue from "helpers/getDisplayedValue";
-import customerHeaders from "helpers/getDisplayedValue/definedHeaders/customerHeaders";
 
 import ICustomer from "interfaces/Customer";
 import {
@@ -17,6 +16,8 @@ import {
   ICreateCustomer,
   IGetCustomer,
 } from "interfaces/customer.service";
+
+import { customerHeaders } from "config/table/headers";
 
 import "./CustomerEdit.scss";
